Add creator validation

+++ src/controllers/movieController.js
@@ -64,6 +64,10 @@ movieController.post("/:movieId/attach", async (req, res) => {
 movieController.get("/:movideId/delete", isAuth, async (req, res) => {
   const movieId = req.params.movideId;
 
+  const movie = await movieService.getOne(movieId);
+  if (!movie.creator?.equals(req.user.id)) {
+    return res.redirect("/");
+  }
   await movieService.delete(movieId);
 
   res.redirect("/");
